requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
accelerate==1.2.1
datasets==3.2.0
nltk==3.9.1
numpy==1.26.4
peft==0.14.0
rouge-score==0.1.2
sacrebleu==2.4.3
sentence-transformers==3.4.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0.dev0
unbabel-comet==2.2.4
wandb==0.19.1

bert-score==0.3.13
joblib==1.2.0
